requirements.txt
langgraph
langchain-google-genai
langchain-core
fastapi
uvicorn
python-dotenv
pydantic
streamlit
httpx
numpy>=1.26.0
grpcio
